+++ FO-DICOM.Core/IO/DirectoryReference.cs
@@ -23,7 +23,7 @@ namespace FellowOakDicom.IO
         #region CONSTRUCTORS
 
         /// <summary>
-        /// Initializes a <see cref="DesktopDirectoryReference"/> object.
+        /// Initializes a <see cref="DirectoryReference"/> object.
         /// </summary>
         /// <param name="directoryName">Name of the directory.</param>
         public DirectoryReference(string directoryName)