+++ FO-DICOM.Core/Network/DesktopNetworkStream.cs
@@ -185,9 +185,9 @@ namespace FellowOakDicom.Network
         #region METHODS
 
         /// <summary>
-        /// Get corresponding <see cref="Stream"/> object.
+        /// Get corresponding <see cref="System.IO.Stream"/> object.
         /// </summary>
-        /// <returns>Network stream as <see cref="Stream"/> object.</returns>
+        /// <returns>Network stream as <see cref="System.IO.Stream"/> object.</returns>
         public Stream AsStream()
         {
             return _networkStream;