+++ FO-DICOM.Core/Network/INetworkListener.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) 2012-2021 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
+using Microsoft.Extensions.Logging;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -15,7 +16,7 @@ namespace FellowOakDicom.Network
         /// <summary>
         /// Start listening.
         /// </summary>
-        /// <returns>An await:able <see cref="Task"/>.</returns>
+        /// <returns>An awaitable <see cref="System.Threading.Tasks.Task"/>.</returns>
         Task StartAsync();
         /// <summary>
         /// Stop listening.
@@ -29,8 +30,9 @@ namespace FellowOakDicom.Network
         /// <param name="noDelay">No delay?</param>
         /// <param name="receiveBufferSize">The size of the receive buffer of the underlying TCP connection</param>
         /// <param name="sendBufferSize">The size of the send buffer of the underlying TCP connection</param>
+        /// <param name="logger">The logger</param>
         /// <param name="token">Cancellation token.</param>
         /// <returns>Connected network stream.</returns>
-        Task<INetworkStream> AcceptNetworkStreamAsync(string certificateName, bool noDelay, int? receiveBufferSize, int? sendBufferSize, CancellationToken token);
+        Task<INetworkStream> AcceptNetworkStreamAsync(string certificateName, bool noDelay, int? receiveBufferSize, int? sendBufferSize, ILogger logger, CancellationToken token);
     }
 }