+++ Tests/FO-DICOM.Tests/Network/AsyncDicomCEchoProviderTests.cs
@@ -94,6 +94,7 @@ namespace FellowOakDicom.Tests.Network
 
         public async Task<DicomCEchoResponse> OnCEchoRequestAsync(DicomCEchoRequest request)
         {
+            await Task.Yield();
             return new DicomCEchoResponse(request, DicomStatus.Success);
         }
     }