+++ Tests/FO-DICOM.Tests/Network/AsyncDicomCGetProviderTests.cs
@@ -130,6 +130,7 @@ namespace FellowOakDicom.Tests.Network
 
         public async IAsyncEnumerable<DicomCGetResponse> OnCGetRequestAsync(DicomCGetRequest request)
         {
+            await Task.Yield();
             yield return new DicomCGetResponse(request, DicomStatus.Success);
         }
 
@@ -173,6 +174,7 @@ namespace FellowOakDicom.Tests.Network
 
         public async IAsyncEnumerable<DicomCGetResponse> OnCGetRequestAsync(DicomCGetRequest request)
         {
+            await Task.Yield();
             yield return new DicomCGetResponse(request, DicomStatus.Pending);
             yield return new DicomCGetResponse(request, DicomStatus.Pending);
             yield return new DicomCGetResponse(request, DicomStatus.Success);