+++ Tests/FO-DICOM.Tests/Network/Client/DicomClientTest.cs
@@ -20,6 +20,9 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
+// These tests cover some obsolete properties
+#pragma warning disable CS0618
+
 namespace FellowOakDicom.Tests.Network.Client
 {
     [Collection("Network"), Trait("Category", "Network")]